+++ R/helper_assess.R
@@ -1017,10 +1017,10 @@ calculateRhoT <- function(
 #' HTMT
 #'
 #' Computes either the heterotrait-monotrait ratio of correlations (HTMT) based on 
-#' \insertCite{Henseler2015;textual}{cSEM} or its advancement HTMT2. While the HTMT
-#' is a consistent estimator for the construct correlation in case of tau-equivalent
-#'  measurement models, the HTMT2 is a consistent estimator for congeneric measurement
-#'  models. In general, they are used to assess discriminant validity.
+#' \insertCite{Henseler2015;textual}{cSEM} or its advancement HTMT2 proposed by \insertCite{Roemerinprint;textual}{cSEM}.
+#' While the HTMT is a consistent estimator for the construct correlation in 
+#' case of tau-equivalent measurement models, the HTMT2 is a consistent estimator
+#' for congeneric measurement models. In general, they are used to assess discriminant validity.
 #' 
 #' Computation of the HTMT assumes that all intra-block and inter-block 
 #' correlations between indicators are either all-positive or all-negative.
@@ -1032,7 +1032,7 @@ calculateRhoT <- function(
 #' to compute the 1-alpha%-quantile, use `.ci`. To control the bootstrap process,
 #' arguments `.handle_inadmissibles`, `.R` and `.seed` are available. 
 #' 
-#' Since the HTMT is defined with respect to a classical true score measurement
+#' Since the HTMT and the HTMT2 both assume a reflective measurement
 #' model only concepts modeled as common factors are considered by default.
 #' For concepts modeled as composites the HTMT may be computed by setting
 #' `.only_common_factors = FALSE`, however, it is unclear how to
@@ -1178,11 +1178,26 @@ calculateHTMT <- function(
     }else{
       monocor2<-monocortemp[lower.tri(monocortemp)]
     }
-    hetcor<-c(S[x[[1]],x[[2]]])
+    # take always the absolute value of the heterotrait-heteromethod correlations
+    hetcor<-abs(c(S[x[[1]],x[[2]]]))
     
     # return correlations as list of vectors containing correlations
     list(monocor1,monocor2,hetcor)
   })
+  
+  # check if sign of all heterotrait-heteromethod correlations is negative; 
+  # if this i the case -1 else 1
+  sign_identification=sapply(block_pairs,function(x){
+    S_signs_two_blocks=S_signs[names(x)[1],names(x)[2]]
+    S_elements_two_blocks=S_elements[names(x)[1],names(x)[2]]
+    
+    if(S_signs_two_blocks==S_elements_two_blocks){
+      1
+    }else{
+      -1
+    }
+  })
+  
   
   if(.type_htmt=='htmt2'){
     # calculate geometric mean of the correlations
@@ -1206,6 +1221,9 @@ calculateHTMT <- function(
   htmts <- sapply(avg_cor,function(x){
     x[3]/sqrt(x[1]*x[2])
   })
+  
+  # 
+  htmts <- htmts * sign_identification 
   
   # Sort HTMT values in matrix
   out<-matrix(0,