+++ R/zz_datasets.R
@@ -137,7 +137,7 @@
 #' a <- csem(threecommonfactors, model_correct)
 #' 
 #' ## The overall model fit is evidently almost perfect:
-#' testOMF(a, .R = 30, .verbose = FALSE) # .R = 30 to speed up the example
+#' testOMF(a, .R = 30) # .R = 30 to speed up the example
 "threecommonfactors"
 
 #' Data: Second order common factor of composites